pyproject.toml
[project]
name = "terna-etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apache-airflow>=3.1.8",
    "apache-airflow-providers-postgres>=6.6.2",
    "feedparser>=6.0.12",
    "polars>=1.39.3",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
